improved version of the fish population program
The first line imports the pyplot module from the matplotlib library, which we will use to plot the fish population data.

The calculate_fish_population function takes three arguments: initial_fish_population, carrying_capacity, and growth_rate. It calculates and returns the fish population at each year based on these values using the formula provided in the original program.

The growth rate r is set to 0.05.

The user is prompted to input the initial fish population, carrying capacity, and number of years to be modelled. These values are stored in the variables initial_fish_population, carrying_capacity, and years, respectively.

The input values are validated to make sure they are all positive. If any of the values is negative, an error message is displayed and the program exits.

An empty list called fish_populations is created to store the fish population for each year.

A for loop iterates over the range of years from 1 to years + 1. For each iteration, the fish population for that year is calculated using the calculate_fish_population function and the previous year's fish population as the initial population. The result is appended to the fish_populations list.

The results are printed to the console, including the year and the number of fish for each year.

The plot function from pyplot is used to plot the fish population data stored in the fish_populations list. The xlabel, ylabel, and title functions are used to add labels to the x-axis, y-axis, and plot title, respectively. Finally, the show function is called to display the plot.